requirements.txt
mercantile==1.2.1
Pillow==9.0.1
Flask==2.0.2
flask-restx==1.1.0
